Replace all uses of `Path` with `CString`.
All the `Path`s were being converted to `CString`s on every call to any
function that accepted them, which is an unnecessary allocation (and the
possibility of a panic which should probably happen outside the
functions in this crate) which can be easily avoided now.

You can fix existing code by converting `Path`s to `CString` using
something like:

    CString::from_slice(path.container_as_bytes())

and then passing that to the function(s).

Or just not create `Path`s and pass a `&[u8]` to `CString::from_slice`
directly.

    foo(&Path::new("..."))

to

    foo(&CString::from_slice(b"..."))

[breaking-change]