ioctl: modify type signatures to match common ioctl usage and libc impls
The ioctl interface expected by most external parties is that of
an unsigned long (c_ulong).  The libc crate expects as an int.  This
change exposes c_ulong (via ioctl_op_t) to external users and
casts to c_int before calling the libc function.

The only documentation in SUSv3 for ioctl has to do with stream
devices.  The support documented there is not widely supported (it
is not supported by the Linux kernel)[1].

[1] The Linux Programming Interface, Page 86.