Fix AsExtStr to return CStrings
As described in #117, the AsExtStr trait is defined to return a
raw *const libc::c_char. Its impl for OsStr simply borrowed the
byte slice from its OsStr argument and cast it to a
*const libc::c_char, which does not construct a proper null-terminated C
string.

This changes the AsExtStr trait to return a Result<CString>, and changes
the impl for OsStr to allocate a new owned CString argument from its
OsStr argument and return that instead.

Fixes #117
Fixes #120