Fix NixPath yield with CStr instead of OsStr
As described in #117, the `AsExtStr` trait is defined to return a raw `*const
libc::c_char`. Its impl for `OsStr` simply borrowed the byte slice from its
`OsStr` argument and cast it to a `*const libc::c_char`, which does not
construct a proper null-terminated C string.

Given this, the `AsExtStr` is not necessary and is removed. `NixPath` is
updated to yield `CStr`.

Fixes #117, #120
Thanks to @dead10ck