[sys::signal] implement signal masking
This patch provides an ffi funcion wrapper for `pthread_sigmask`,
which can be used for blocking or unblocking signals for the calling
thread.

`pthread_sigmask` has the same interface as the `sigprocmask`
syscall, however calling `sigprocmask` in a multithreaded program is
undefined behavior. Because the Rust stdlib links `pthreads` as a
dependency, therefore making every Rust program multithreaded,
`sigprocmask` is not implemented in this patch.