libc: re-export libc properly
With rust 1.7, the following warning was being emitted by the compiler:

    warning: `pub extern crate` does not work as expected and should
    not be used. Likely to become an error. Prefer `extern crate` and
    `pub use`.

Based on https://github.com/rust-lang/rust/issues/26775 it appears that
the warning in 1.7 which was to be escalated to an error is going away
but in older versions of rust it is still the case that `pub extern crate`
will not work as expected.  Instead, we use a somewhat creative hack to
export the libc root as a module in nix.  Down the line, it may make
sense to either eliminate the need to export libc (by chaning the ioctl
macros) or to move toward deprecated older versions of rustc.

Signed-off-by: Paul Osborne <osbpau@gmail.com>