testing: add ci scripts to aid testing on more architectures/versions
These scripts represent a rough first cust at adding support for
locally building nix and its tests for a large number of rust
versions and architectures.  Currently, this mostly helps with
testing that nix compiles for various architectures (to the point that
there are a mound of problems that need to be addressed).  Once these
are sorted, the code for actually running the tests for various
architectures will be approached (this will require qemu and will
still not target all architectures).

Signed-off-by: Paul Osborne <osbpau@gmail.com>