testing: docker tests now work on travis
All of the platforms tested currently fail in some way currently, but
the infrastructure itself appears to be working.  As we have moved to
the legacy infrastructure (required in order to use docker), the tests
do run slower now.  To compensate for that, i686 builds are only done on
stable and we do not test every version between the current stable and
the oldest version we support.

Signed-off-by: Paul Osborne <osbpau@gmail.com>