testing: ensure run.sh is executing in docker container
On the host, some undesired things could end up happening if this
is executed, so we now exit and log if that is attempted.  In addition,
we now backup the previous cargo config if present.  This will never
be the case in the current build setup, but that could change if
we decide to do builds in a data volume in the future.

Signed-off-by: Paul Osborne <osbpau@gmail.com>