ci: push docs to gh-pages and code cov to coveralls
This moves things around a fair bit in the .travis.yml to be
more logically ordered and adds support for publishing docs
(on master) to gh-pages.  Code coverage is also calculated
and published to coveralls.io.

Travis is able to write to gh-pages on the github repository
via the GH_TOKEN environment variable stored in the travis
settings (marked as private).  This token is generated from
the github user 'rust-nix-bot' in order to avoid given access
to other repos associated with an actual user's account (this
practice is recommended by Github).

Other changes worth noting:
 - We now test fewer intermediate versions of rust
 - The existing logic to publish to Carl's S3 bucket have been removed.

Signed-off-by: Paul Osborne <osbpau@gmail.com>