ci: drop official support for rustc < 1.7.0
Supporting older verisons of rust is causing CI to fail and is
a somewhat constant support burden with questionable value.  1.7.0
is the oldest I have seen that people have had a requirement on
supporting in recent history.

It may make sense to work toward a more official policy on
what version of rust we will support in the future.  Users of
older version of rust are, of course, welcome to use older versions
of nix and everything else in the ecosystem.

Signed-off-by: Paul Osborne <osbpau@gmail.com>