unistd: add docs for gethostname and tweak API
Although the underlying C API does take a pointer to a set of
characters, it is a requirement of almost every operating system
that these bytes not contain a premature NUL character or other
special characters.  In other words, you want a `&str`.  Changing
this to make the API make a bit more sense.

Signed-off-by: Paul Osborne <osbpau@gmail.com>