unistd: document and change implementation of gethostname
Previously gethostname just mutated a buffer.  We now provide
a slightly more usable (but still allocation free) API that ensures
that the returned buffer is NUL-terminated.  We give back a `&CStr`
instead of requiring that the user do all of the conversions from
`&[u8]` when we know we are dealing with a `&CStr`.

Signed-off-by: Paul Osborne <osbpau@gmail.com>