signal: impl From<Signal> for SigSet
The normal libc::sigset_t API is incredibly clunky when initialising
SigSets. Since a SigSet semantically is a (mathematical) set of Signals,
it makes sense to allow Signals to be promoted to a SigSet. This will
also allow us to change our API to take Into<SigSet> rather than SigSet,
making this even more idiomatic.

  let mut sigset = SigSet::empty();
  sigset.add(Signal::SIGINT);
  // versus
  let sigset: SigSet = Signal::SIGINT.into();
  let sigset = SigSet::from(Signal::SIGINT);

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>