Fix kevent for netbsd filter datatype
The datatype for kevent.filter is u32 on NetBSD and
i16 on all other supported platforms. This was recently
fixed in upstream libc, breaking this API, so this
fixes it.

This change also modernizes the code a bit to unify the
EventFilter datatype across platforms and switch to the
libc_enum! macro.