src/sys/socket/sockopt.rs: implement bytes type as Vec<u8>;
On macOS / iOS, to get the IFNAME of a UTUN, a `getsockopt(fd, SYSPROTO_CONTROL, UTUN_OPT_IFNAME, val, len);` call must be made passing not a struct, but a CString of `IFNAMSIZ` bytes, which returns how many bytes of the CString was written to (including with null-terminating byte).

This implements getsockopt for bytes.