vendor the bitflags crate
Bitflags has a policy of changing the MSRV in minor releases.  This is a
breaking change, which semver does not allow in minor releases.  Nix
adjusts by locking bitflags to a specific minor version.  But Cargo does
not allow multiple semver-compatible versions of a crate to be included
in the same build, with the result that crates which use both Nix 0.23
and also use bitflags 1.2 cannot compile.

Fix the problem by vendoring the bitflags crate, removing the external
dependency completely.