return None if user does not exists
getpwuid_r and getpwnam_r does return ENOENT if a user does not exists. This might lead to the unexpected case where the outer Result may fail even if only the user does not exist.

This commit changes this behavior to return a successful Result but with an empty Option. It unifies the behavior with older systems.