Don't deref uninit addrs in `pack_mhdr_to_receive`
This adds a new method to `SockaddrLikePriv` to get a pointer to some
underlying `libc::sockaddr` from a pointer to some type. This avoids
dereferencing uninitialized values during `pack_mhdr_to_receive`, which
is called from `MultiHeaders::preallocate` and `recvmsg`.

Because this is a private trait, this is not a semver-breaking change.