Revert "Merge #2027 #2057"
This reverts commit 0c3afc27d7bdbea7e5c12969e1ba0607c14b1dc9.

This rolls back the version of bitflags used to 1.1.  Upgrading bitflags
to 2.0 inadvertently caused breaking changes in the r0.26 branch, by
virtue of the methods that the bitflags! macro adds.

Also, mask the bad_bit_mask lint, triggered by bitflags's generated
code.  That lint was the motivation for upgrading bitflags in the first
place.  The bitflags maintainers have decided not to fix those warnings
in the 1.x release series.
https://github.com/bitflags/bitflags/pull/373

Fixes #2112