2075: Add ControlMessage::Ipv6HopLimit (#2074)
When sending IPv6 packets with `sendmsg`, Linux does not use the hop limit
set on the socket. Instead, the hop limit has to be specified for each
individual message with ancillary data in a cmsg.

This commit adds the enum variant
`ControlMessage::Ipv6HopLimit` to specify the limit. The variant is
available on the `net` feature flag for Linux, MacOs, FreeBSD,
DragonflyBSD, Android, iOS and Haiku.

Co-authored-by: Simon B. Gasse <sgasse@users.noreply.github.com>