Add `signal::sigaction_current()`
Provide a way to query the currently installed sigaction.

The decision to add `sigaction_current()` instead of just exposing the
`sigaction_inner()` function was to avoid any confusion over the
semantics of passing in a `None` `sigaction` argument (eg: someone
thinking that it meant remove or reset the action).

This builds towards #2172.