Correct a comment (#2282)
And while here, make cmsg_space a const fn, though I doubt it will make
a difference since it's already inline.

The original comment was _almost_ correct.  Changing the cmsg_space
macro to return an array works in most locations.  However, it fails in
the test_recverr_impl function in the tests due to a E0401 Error "Inner
items do not inherit the generic parameters from the items they are
embedded in."  That error is very difficult to fix in our test, and it
might occur in user code, too.  The root of the problem is that each
length of array is its own distinct type, and in a generic function
cmsg_space returns an array type that may depend on the generic
parameters.

Alternatively we could provide a second macro that always heap
allocates.  But that may be too complicated just to save a single
allocation.