Fix undefined memory read and associated crash.
Linux does set the amaster parameter of `forkpty()` to a valid file
descriptor in the child, but we pass this value to
`OwnedFd::from_raw_fd`, which panics if the passed value is
`0xffff_ffff`.

A simple reproducer:

```
fn effify()
{
    let x: [u8; 1024] = [0xff; 1024];
}

fn main()
{
    effify();

    let dupped = nix::unistd::dup(1).unwrap();
    let result = unsafe { nix::pty::forkpty(None, None).unwrap() };

    match result.fork_result
    {
        nix::unistd::ForkResult::Child =>
        {
            nix::unistd::dup2(dupped, 1).unwrap();
            println!("Child is happy");
        }
        nix::unistd::ForkResult::Parent { child } =>
        {
            println!("Parent is happy");
            println!("{:?}", nix::sys::wait::waitpid(child, None).unwrap());
        }
    }
}
```

The child process panics in the above code without the above patch
applied.