Fix syscall_info (#2653)
According to the man page:

> The addr argument contains the size of the buffer pointed to by the
> data argument (i.e., sizeof(struct ptrace_syscall_info)).

We were setting the data argument, so this syscall was returning
garbage. This is easily reproduced and verified by looking at the `op`
value, which should range from 0 to 3. The PR associated with this
commit contains a sample program to demonstrate this.

The fix is done in the `ptrace_get_data` helper to avoid duplicating its
implementation just for `syscall_info`.

Of all the other callers, all but one are documented as ignoring `addr`.
The other one (`PTRACE_GETREGS`) is documented as ignoring `addr`
_except on SPARC systems_, on which `addr` and `data` are reversed.
However, this is already not respected by `nix`, so this changes is not
disruptive in this regard.

There should be no performance concerns as we are replacing one constant
with another.

Co-authored-by: Cédric Barreteau <cbarrete@users.noreply.github.com>