Skip lio_listio doc test on musl/aarch64/i686 as it hangs
Do not run this doc test on:

* aarch64-unknown-linux-musl
* i686-unknown-linux-musl

because it hangs on these targets. After further debugging, we think this is
likely a bug of musl. Since we only test our bindings and do not intend to
fix the underlying libc bug, we skip this test here. See this thread [1] for
the discussion of this issue

[1]: https://github.com/nix-rust/nix/pull/2689#issuecomment-3419813159