Fix CI by relaxing our MSRV constraint (#2689)
* Fix CI by relaxing our MSRV constraint

Recently a transitive dev-dependency published a new version that is
incompatible with Nix's MSRV, causing CI to break.  That's annoying,
because there's really no reason why a crate's dev-dependencies ought to
respect the MSRV.  Fix Nix's CI by:

* Running all tests with stable Rust (or nightly, for certain targets)
* Adding an additional set of CI checks to ensure that the crate will
  compile with MSRV, on all of the major operating systems, but don't
  try to compile its tests.
* Eliminate the now-redundant "rust-stable" CI task.

* Don't use --all-targets when testing MSRV

* Fix some Clippy warnings on apple

I think these never showed up in CI before, because we've never run
clippy on Rust stable on apple until now.

* Suppress a Clippy false positive

* GetCString only needs to be build with features=net

* Fix Clippy on Solarish, with Rust stable

* Disable the libc::personality doc test on QEMU

It's failing there with EPERM.  It's probably seccomp's fault

* Skip lio_listio doc test on musl/aarch64/i686 as it hangs

Do not run this doc test on:

* aarch64-unknown-linux-musl
* i686-unknown-linux-musl

because it hangs on these targets. After further debugging, we think this is
likely a bug of musl. Since we only test our bindings and do not intend to
fix the underlying libc bug, we skip this test here. See this thread [1] for
the discussion of this issue

[1]: https://github.com/nix-rust/nix/pull/2689#issuecomment-3419813159

* Run rustfmt on src/sys/aio.rs

* Set RUST_BACKTRACE=1 when running tests

Set this to enable the backtrace to make debugging easier

* Cross env var passthrough: RUSTDOCFLAGS and RUST_BACKTRACE

I don't know if "RUSTDOCFLAGS" will be passed automatically, but from
their doc [1]:

> In the instances that you do want to pass through additional environment
> variables, this can be done via build.env.passthrough in your Cross.toml:
>
> ```toml
> [build.env]
> passthrough = [
>    "RUST_BACKTRACE"
> ]
> ```

We do need to specify "RUST_BACKTRACE" here. Anyway, let's pass both in
case Cross won't do this for us.

[1]: https://github.com/cross-rs/cross/blob/49cd054de9b832dfc11a4895c72b0aef533b5c6a/docs/environment_variables.md#environment-variable-passthrough

* Disable the ScmTimestamp doc test on 32-bit musl

Because this platform has very unique behavior.  It can be worked
around, but our doc test does not include that workaround.

https://github.com/nix-rust/nix/issues/2698

---------

Co-authored-by: Steve Lau <stevelauc@outlook.com>